The game
Added new systems such as debuff icons, major debugging on existing ones.
It's pretty much at the absolute minimum I was expecting to submit. It'd be a lot more interesting with some more debuffs, but more importantly it's super ugly. I'm gonna be crunching on this, shouldn't have waited this long.